# fix(db): Improve query performance when deleting tokens for public clients