Fixes https://github.com/mozilla/fxa-auth-server/issues/2695

As noted in https://github.com/mozilla/fxa-auth-server/issues/2695, the performance of this query was causing problems in production:

```
DELETE FROM tokens WHERE userId=? AND clientId IN (?);
```

It's fast if MySQL decides to use the `tokens_user_id` index, and it's very slow of if MySQL decides to use the `tokens_client_id` index.  Unfortunately MySQL's choice here doesn't seem very scrutable - in my testing it will choose `tokens_client_id` if the list of clients has only a single element, while in production it appears to choose `tokens_client_id` unless there's a foreign key constraint on the `clientId` column.

Double unfortunately, we can't use index hints to influence MySQL's choice, because they're not supported on raw delete statements.

What we *can* do is switch to MySQL's multi-table delete syntax, which appears to allow index hints in the join, and which has the added advantage of saving us a query to read the list of public clients.

Here's the `EXPLAIN` output on my test db, which has no foreign-key constraints on the `tokens` table.  First for the original query which doesn't consistently use the right index:

```
mysql> EXPLAIN DELETE FROM tokens WHERE userId=UNHEX('08C263A41CDC39C90F609F2A9FA12EF3') AND clientId IN (UNHEX('00A8F8482522BE5B'));
+----+-------------+--------+------------+-------+---------------------------------+------------------+---------+-------+------+----------+-------------+
| id | select_type | table  | partitions | type  | possible_keys                   | key              | key_len | ref   | rows | filtered | Extra       |
+----+-------------+--------+------------+-------+---------------------------------+------------------+---------+-------+------+----------+-------------+
|  1 | DELETE      | tokens | NULL       | range | tokens_client_id,tokens_user_id | tokens_client_id | 8       | const |    1 |   100.00 | Using where |
+----+-------------+--------+------------+-------+---------------------------------+------------------+---------+-------+------+----------+-------------+
1 row in set (0.00 sec)

mysql> EXPLAIN DELETE FROM tokens WHERE userId=UNHEX('08C263A41CDC39C90F609F2A9FA12EF3') AND clientId IN (UNHEX('00A8F8482522BE5B'), UNHEX('FB88B057B5C0AED8'));
+----+-------------+--------+------------+-------+---------------------------------+----------------+---------+-------+------+----------+-------------+
| id | select_type | table  | partitions | type  | possible_keys                   | key            | key_len | ref   | rows | filtered | Extra       |
+----+-------------+--------+------------+-------+---------------------------------+----------------+---------+-------+------+----------+-------------+
|  1 | DELETE      | tokens | NULL       | range | tokens_client_id,tokens_user_id | tokens_user_id | 16      | const |    1 |   100.00 | Using where |
+----+-------------+--------+------------+-------+---------------------------------+----------------+---------+-------+------+----------+-------------+
1 row in set (0.00 sec)
```

And for the modified query:

```
mysql> EXPLAIN DELETE tokens FROM tokens FORCE INDEX (tokens_user_id) INNER JOIN clients ON tokens.clientId = clients.id WHERE tokens.userId=UNHEX('08C263A41CDC39C90F609F2A9FA12EF3') AND (clients.publicClient = 1 OR clients.canGrant = 1);
+----+-------------+---------+------------+--------+----------------+----------------+---------+---------------------------+------+----------+-------------+
| id | select_type | table   | partitions | type   | possible_keys  | key            | key_len | ref                       | rows | filtered | Extra       |
+----+-------------+---------+------------+--------+----------------+----------------+---------+---------------------------+------+----------+-------------+
|  1 | DELETE      | tokens  | NULL       | ref    | tokens_user_id | tokens_user_id | 16      | const                     |    1 |   100.00 | Using where |
|  1 | SIMPLE      | clients | NULL       | eq_ref | PRIMARY        | PRIMARY        | 8       | fxa_oauth.tokens.clientId |    1 |    19.00 | Using where |
+----+-------------+---------+------------+--------+----------------+----------------+---------+---------------------------+------+----------+-------------+
2 rows in set (0.00 sec)
```

The `FORCE INDEX (tokens_user_id) ` doesn't appear to be necessary here, as I get the same result without it:

```
mysql> EXPLAIN DELETE tokens FROM tokens INNER JOIN clients ON tokens.clientId = clients.id WHERE tokens.userId=UNHEX('08C263A41CDC39C90F609F2A9FA12EF3') AND (clients.publicClient = 1 OR clients.canGrant = 1);
+----+-------------+---------+------------+--------+---------------------------------+----------------+---------+---------------------------+------+----------+-------------+
| id | select_type | table   | partitions | type   | possible_keys                   | key            | key_len | ref                       | rows | filtered | Extra       |
+----+-------------+---------+------------+--------+---------------------------------+----------------+---------+---------------------------+------+----------+-------------+
|  1 | DELETE      | tokens  | NULL       | ref    | tokens_client_id,tokens_user_id | tokens_user_id | 16      | const                     |    1 |   100.00 | Using where |
|  1 | SIMPLE      | clients | NULL       | eq_ref | PRIMARY                         | PRIMARY        | 8       | fxa_oauth.tokens.clientId |    1 |    19.00 | Using where |
+----+-------------+---------+------------+--------+---------------------------------+----------------+---------+---------------------------+------+----------+-------------+
2 rows in set (0.00 sec)
```

But I don't think it can hurt, so it seemed prudent to include it.

@jrgm, when you get a chance, could we please try an `EXPLAIN` of these modified queries in stage or prod to check the query plan there?

If we like this solution, I'll add a migration that re-drops the foreign key constraints that we re-introuces in https://github.com/mozilla/fxa-oauth-server/pull/623.  I havent added it yet because we need to get that merged and ported over to this repo.