@jrgm 2018-11-01
I picked a uid at semi-random in stage and I get these explains:
```
EXPLAIN EXTENDED DELETE FROM tokens WHERE userId=UNHEX('E0791B95947B44E2A69F4FD21013FBAE') AND clientId IN (UNHEX('EA3CA969F8C6BB0D'));     
+----+-------------+--------+-------+---------------------------------+----------------+---------+-------+------+----------+-------------+  
| id | select_type | table  | type  | possible_keys                   | key            | key_len | ref   | rows | filtered | Extra       |  
+----+-------------+--------+-------+---------------------------------+----------------+---------+-------+------+----------+-------------+  
|  1 | SIMPLE      | tokens | range | tokens_client_id,tokens_user_id | tokens_user_id | 16      | const | 2176 |   100.00 | Using where |  
+----+-------------+--------+-------+---------------------------------+----------------+---------+-------+------+----------+-------------+  
                                                                                                                                            
EXPLAIN EXTENDED DELETE FROM tokens WHERE userId=UNHEX('E0791B95947B44E2A69F4FD21013FBAE') AND clientId IN (UNHEX('EA3CA969F8C6BB0D'), UNHEX('5882386C6D801776'));                                                                                                                     
+----+-------------+--------+-------+---------------------------------+----------------+---------+-------+------+----------+-------------+  
| id | select_type | table  | type  | possible_keys                   | key            | key_len | ref   | rows | filtered | Extra       |  
+----+-------------+--------+-------+---------------------------------+----------------+---------+-------+------+----------+-------------+  
|  1 | SIMPLE      | tokens | range | tokens_client_id,tokens_user_id | tokens_user_id | 16      | const | 2176 |   100.00 | Using where |  
+----+-------------+--------+-------+---------------------------------+----------------+---------+-------+------+----------+-------------+
```
```
EXPLAIN EXTENDED DELETE tokens FROM tokens FORCE INDEX (tokens_user_id) INNER JOIN clients ON tokens.clientId = clients.id WHERE tokens.userId=UNHEX('E0791B95947B44E2A69F4FD21013FBAE') AND (clients.publicClient = 1 OR clients.canGrant = 1);
+----+-------------+---------+--------+----------------+----------------+---------+---------------------------+------+----------+-------------+
| id | select_type | table   | type   | possible_keys  | key            | key_len | ref                       | rows | filtered | Extra       |
+----+-------------+---------+--------+----------------+----------------+---------+---------------------------+------+----------+-------------+
|  1 | SIMPLE      | tokens  | ref    | tokens_user_id | tokens_user_id | 16      | const                     | 2176 |   100.00 | Using where |
|  1 | SIMPLE      | clients | eq_ref | PRIMARY        | PRIMARY        | 8       | fxa_oauth.tokens.clientId |    1 |   100.00 | Using where |
+----+-------------+---------+--------+----------------+----------------+---------+---------------------------+------+----------+-------------+

EXPLAIN DELETE tokens FROM tokens INNER JOIN clients ON tokens.clientId = clients.id WHERE tokens.userId=UNHEX('E0791B95947B44E2A69F4FD21013FBAE') AND (clients.publicClient = 1 OR clients.canGrant = 1);
+----+-------------+---------+--------+---------------------------------+----------------+---------+---------------------------+------+-------------+
| id | select_type | table   | type   | possible_keys                   | key            | key_len | ref                       | rows | Extra       |
+----+-------------+---------+--------+---------------------------------+----------------+---------+---------------------------+------+-------------+
|  1 | SIMPLE      | tokens  | ref    | tokens_client_id,tokens_user_id | tokens_user_id | 16      | const                     | 2176 | Using where |
|  1 | SIMPLE      | clients | eq_ref | PRIMARY                         | PRIMARY        | 8       | fxa_oauth.tokens.clientId |    1 | Using where |
+----+-------------+---------+--------+---------------------------------+----------------+---------+---------------------------+------+-------------+
```

Those were run with the FK still in place. (I can run again with the FK dropped).

`2176` is close to the actual number of tokens for this `uid`. I don't know why this random user has that many tokens in stage.

Now all of those tokens are expired. I tried adding `tokens.expiredAt > NOW()` on the `WHERE`, but the explain did not improve. I'm not sure why the index on expiresAt is not helping here.

This pruning really needs to happen asap, and that's my first priority in the/my morning.