@rfk 2018-11-12
Taking the WIP off of this, I think it's ready to go.  It modifies the badly-performing queries and adds a new migration to drop the foreign key constraints.  (IIUC we can apply this migration without using gh-ost, so we might be willing to try it in the train-125 deploy; @jrgm thoughts?)